blueprint update
