+++ __init__.py
@@ -1,0 +1,2 @@
+# This file makes the root directory a Python package
+# This helps with imports
