+++ blueprints/__init__.py
@@ -1,0 +1,1 @@
+# Empty __init__.py to make the directory a Python package
