+++ blueprints/api/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+api_bp = Blueprint('api', __name__)
+
+from . import routes
