+++ blueprints/data/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+data_bp = Blueprint('data', __name__, template_folder='templates')
+
+from . import routes
