+++ blueprints/main/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+main_bp = Blueprint('main', __name__)
+
+from . import routes
