+++ blueprints/profile/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+profile_bp = Blueprint('profile', __name__)
+
+from . import routes
