+++ templates/debug_file.html
@@ -61,7 +61,7 @@ endblock %} {% block content %}
   </div>
 
   <div class="mt-4">
-    <a href="{{ url_for('dashboard') }}" class="btn btn-secondary"
+    <a href="{{ url_for('main.dashboard') }}" class="btn btn-secondary"
       >Back to Dashboard</a
     >
   </div>
