+++ templates/edit_profile.html
@@ -60,7 +60,7 @@
       <button type="submit" class="btn btn-primary">
         <i class="bi bi-save me-1"></i> Save Changes
       </button>
-      <a href="{{ url_for('profile') }}" class="btn btn-secondary ms-2">Cancel</a>
+      <a href="{{ url_for('profile.profile') }}" class="btn btn-secondary ms-2">Cancel</a>
     </div>
   </form>
 </div>
