+++ templates/forgot_password.html
@@ -8,7 +8,7 @@
         <div class="card-body">
             <h3 class="card-title text-center mb-4">Forgot Your Password?</h3>
 
-            <form method="POST" action="{{ url_for('forgot_password') }}">
+            <form method="POST" action="{{ url_for('auth.forgot_password') }}">
                 <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
                 
                 <div class="mb-3">
