+++ templates/login.html
@@ -13,7 +13,7 @@
         <div class="card-body">
             <h3 class="card-title text-center mb-4">Sign In</h3>
 
-            <form method="POST" action="{{ url_for('login') }}">
+            <form method="POST" action="{{ url_for('auth.login') }}">
                 <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
                 
                 <div class="mb-3">
@@ -44,7 +44,7 @@
                 </div>
                 
                 <div class="text-center mt-3">
-                    <a href="{{ url_for('forgot_password') }}" class="forgot-link">
+                    <a href="{{ url_for('auth.forgot_password') }}" class="forgot-link">
                         Forgot your password?
                     </a>
                 </div>
@@ -57,7 +57,7 @@
     </div>
     
     <div class="text-center mt-3">
-        <p>Don't have an account? <a href="{{ url_for('signup') }}" class="text-decoration-none">Sign up</a></p>
+        <p>Don't have an account? <a href="{{ url_for('auth.signup') }}" class="text-decoration-none">Sign up</a></p>
     </div>
 </div>
 {% endblock %}
