+++ templates/profile.html
@@ -19,7 +19,7 @@ content %}
             {{ user.work_location or 'No Work location set' }}
           </p>
           <a
-            href="{{ url_for('edit_profile') }}"
+            href="{{ url_for('profile.edit_profile') }}"
             class="btn btn-primary btn-sm mt-2"
           >
             <i class="bi bi-pencil-square"></i>Edit Profile
