+++ templates/reset_password.html
@@ -8,7 +8,7 @@
         <div class="card-body">
             <h3 class="card-title text-center mb-4">Reset Your Password</h3>
 
-            <form method="POST" action="{{ url_for('reset_password', token=token) }}">
+            <form method="POST" action="{{ url_for('auth.reset_password', token=token) }}">
                 <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
                 
                 <div class="mb-3">
