+++ templates/reset_success.html
@@ -11,7 +11,7 @@
             <p class="lead">You can now log in with your new password.</p>
 
             <div class="d-grid gap-2 mt-4">
-                <a href="{{ url_for('login') }}" class="btn btn-primary">
+                <a href="{{ url_for('auth.login') }}" class="btn btn-primary">
                     Go to Login
                 </a>
             </div>
