+++ templates/visualize.html
@@ -81,7 +81,7 @@
         </div>
         <div class="col-md-4 text-end">
             <div class="btn-group">
-                <a href="{{ url_for('dashboard') }}" class="btn btn-outline-secondary">
+                <a href="{{ url_for('main.dashboard') }}" class="btn btn-outline-secondary">
                     <i class="bi bi-arrow-left"></i> Back to Dashboard
                 </a>
                 <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">
@@ -93,7 +93,7 @@
                 </ul>
             </div>
             <div class="mt-2">
-                <a href="{{ url_for('view_charts', dataset_id=dataset.id) }}" class="btn btn-outline-info w-100">
+                <a href="{{ url_for('data.view_charts', dataset_id=dataset.id) }}" class="btn btn-outline-info w-100">
                     <i class="bi bi-bar-chart-line"></i> View Charts
                 </a>
             </div>
